add TourRoot
