+++ src/packages/tour/tour.ts
@@ -22,7 +22,7 @@ import DOMEvent from "../../util/DOMEvent";
 import onKeyDown from "./onKeyDown";
 import onResize from "./onResize";
 import van from "../dom/van";
-import appendChild from "../../util/appendChild";
+import { TourRoot } from "./tourRoot";
 
 /**
  * Intro.js Tour class
@@ -34,7 +34,6 @@ export class Tour implements Package<TourOptions> {
   private _direction: "forward" | "backward";
   private readonly _targetElement: HTMLElement;
   private _options: TourOptions;
-  private _root: HTMLElement;
 
   private readonly callbacks: {
     beforeChange?: introBeforeChangeCallback;
@@ -377,28 +376,16 @@ export class Tour implements Package<TourOptions> {
     }
   }
 
-  // temporary
-  public getRoot = () => this._root;
-
-  public appendToRoot(element: HTMLElement, animate = false) {
-    appendChild(this._root, element, animate);
-  }
-
   private createRoot() {
-    if (!this._root) {
-      const { div } = van.tags;
-      this._root = div({ className: "introjs" });
-      this.getTargetElement().appendChild(this._root);
-    }
+    van.add(this.getTargetElement(), TourRoot({ tour: this }));
   }
 
   /**
    * Starts the tour and shows the first step
    */
   async start() {
-    this.createRoot();
-
     if (await start(this)) {
+      this.createRoot();
       this.enableKeyboardNavigation();
       this.enableRefreshOnResize();
     }
